% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_helpers.R
\name{font_hoist}
\alias{font_hoist}
\title{Register style variants of a font as their own families}
\usage{
font_hoist(family, silent = FALSE)
}
\arguments{
\item{family}{The font family}

\item{silent}{Whether to print newly registered font families in a message}
}
\description{
Register style variants of a font as their own families
}
\examples{
\dontrun{
font_hoist(family = "Font Awesome 5 Free")
}

}
