% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsave_auto.R
\name{ggsave_auto}
\alias{ggsave_auto}
\title{Accumulate multiple versions of a plot in interactive mode}
\usage{
ggsave_auto(
  folder = "ggsaveauto_img",
  n_pad = 3,
  padding = stringr::str_pad(1, n_pad, pad = "0"),
  img_prefix = "img_",
  img_ext = ".png",
  width = 20,
  height = 12,
  unit = "cm",
  plot_load = "app",
  ...
)
}
\arguments{
\item{folder}{Passed to \code{path} variable in \code{ggsave()}}

\item{n_pad}{Number of \code{"0"} paddings}

\item{padding}{Unique identifier for current plot}

\item{img_prefix}{Prefix of the image file name}

\item{img_ext}{Extension of the iamge file name}

\item{width}{Passed to \code{ggsave()}}

\item{height}{Passed to \code{ggsave()}}

\item{unit}{Passed to \code{ggsave()}}

\item{plot_load}{Options for loading the saved plot for viewing. One of \code{c("app", "rstudio", "none")}.}

\item{...}{Passed to \code{ggsave()}}
}
\value{
Returns the value of the current/new graphic device. You can suppress this with \code{invisible()}
}
\description{
Wrapper around \code{ggplot2::ggsave()}
}
