% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_paste0.R
\name{\%+\%}
\alias{\%+\%}
\title{Adds two strings together in infix form}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{a string}

\item{rhs}{a string}
}
\value{
A string of length 1
}
\description{
Adds two strings together in infix form
}
\examples{
"Hello" \%+\% ", " \%+\% "world!"
}
