% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{icon_html}
\alias{icon_html}
\title{Find a Font Awesome icon in the appropriate html format}
\usage{
icon_html(icon, type = "css")
}
\arguments{
\item{icon}{A character string that matches one or more Font Awesome icon names}

\item{type}{"css" returns the icon wrapped in \code{span}; "class" returns the icon wrapped in \code{i}}
}
\value{
The first icon match in the specified format
}
\description{
Find a Font Awesome icon in the appropriate html format
}
