% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{junebug_palette}
\alias{junebug_palette}
\title{Selects colors from a palette}
\usage{
junebug_palette(palette, n = 3)
}
\arguments{
\item{palette}{Name of palette in junebug_palettes}

\item{n}{Number of colors to return}
}
\value{
A character vector of colors
}
\description{
Selects colors from a palette
}
