% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{junebug_palettes}
\alias{junebug_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
junebug_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}
