% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{label_unit}
\alias{label_unit}
\title{Annotate a single unit text with a unit label}
\usage{
label_unit(x, unit_label = "ms", position = "last")
}
\arguments{
\item{x}{Vector of labels}

\item{unit_label}{A string attached at the end of the label}

\item{position}{Which label the annotation should go to. Either 'first' (bottom/left) or 'last' (top/right).}
}
\description{
For use in the \code{labels} argument of \verb{ggplot2::scale_*_continuous}
}
\examples{
\dontrun{
my_plot <- ggplot(mtcars, aes(mpg, hp)) + geom_point()
my_plot + scale_y_continuous(labels = label_unit)
my_plot + scale_y_continuous(labels = function(x) label_unit(x, "\%"))
my_plot + scale_y_continuous(labels = function(x) label_unit(x, "\%", "first"))
my_plot + scale_y_continuous(labels = function(x) label_unit(x, "sec"))
}
}
