% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{latex_query}
\alias{latex_query}
\title{Generate a CodeCogs URL that renders latex as image}
\usage{
latex_query(
  formula,
  size = "normal",
  dpi = 300,
  transparent = FALSE,
  inline = FALSE
)
}
\arguments{
\item{formula}{An appropriately escaped string of LaTex formula. Using the new string literal syntax \code{r"(...)"} from R 4.0.0 on the existing formula is recommended.}

\item{size}{Font size. One of "tiny", "small", "normal", "large", or "huge"}

\item{dpi}{Image resolution. Defaults to 300}

\item{transparent}{Should the background be transparent? Defaults to \code{FALSE}, which adds a white background.}

\item{inline}{Render as inline formula? If TRUE, results in a more compact image.}
}
\value{
A character string of the URL
}
\description{
Interactive version at https://www.codecogs.com/latex/eqneditor.php
}
\examples{
\dontrun{
latex_query(r"(\beta + \alpha)")
latex_query("\\\\beta + \\\\alpha")
magick::image_read(latex_query(r"(\beta + \alpha)"))
}
}
