% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{one_label}
\alias{one_label}
\title{Annotate a single break on an axis with a label}
\usage{
one_label(x, suffix = "ms", prefix = "", position = "last")
}
\arguments{
\item{x}{Vector of axis breaks}

\item{suffix}{A string attached to the end of the text. Defaults to "ms".}

\item{prefix}{A string attached to the beginning of the text. Defaults to "".}

\item{position}{Which label the annotation should go to. Either 'first' (bottom/left) or 'last' (top/right).}
}
\description{
For use in the \code{labels} argument of \verb{ggplot2::scale_*_continuous}
}
\examples{
\dontrun{
my_plot <- ggplot(mtcars, aes(mpg, hp)) + geom_point()
my_plot + scale_y_continuous(labels = one_label)
my_plot + scale_y_continuous(labels = function(x) one_label(x, "\%"))
my_plot + scale_y_continuous(labels = function(x) one_label(x, "\%", position = "first"))
my_plot + scale_y_continuous(labels = function(x) one_label(x, prefix = "$"))
}
}
