% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsave_auto.R
\name{raggsave_auto}
\alias{raggsave_auto}
\title{Accumulate multiple versions of a plot in interactive mode}
\usage{
raggsave_auto(
  folder = "raggsaveauto_img",
  p = ggplot2::last_plot(),
  n_pad = 3,
  padding = stringr::str_pad(1, n_pad, pad = "0"),
  img_prefix = "img_",
  img_ext = ".png",
  width = 20,
  height = 12,
  unit = "cm",
  res = 300,
  plot_load = "app",
  ...
)
}
\arguments{
\item{folder}{Directory for the plots to be accumulated in}

\item{p}{Plot to be saved. Defaults to last plot printed}

\item{n_pad}{Number of \code{"0"} paddings}

\item{padding}{Unique identifier for current plot}

\item{img_prefix}{Prefix of the image file name}

\item{img_ext}{Extension of the iamge file name}

\item{width}{Passed to \code{ragg_png()}}

\item{height}{Passed to \code{ragg_png()}}

\item{unit}{Passed to \code{ragg_png()}}

\item{res}{Passed to \code{ragg_png()}}

\item{plot_load}{Options for loading the saved plot for viewing. One of \code{c("app", "rstudio", "none")}.}

\item{...}{Passed to \code{ragg_png()}}
}
\value{
Returns the value of the current/new graphic device. You can suppress this with \code{invisible()}
}
\description{
Wrapper around \code{ragg::ragg_png()}
}
