% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibex.R
\name{read.IBEX}
\alias{read.IBEX}
\title{Reads in the \code{results} file from an IBEX experiment as data frame}
\usage{
read.IBEX(results_file)
}
\arguments{
\item{results_file}{The exact text of the \code{results} file as downloaded from IBEX}
}
\value{
A data frame if the controller writes a single line to the results file (e.g., "DashedSentence") OR a list of data frames if the controller writes more than one line to the results file (e.g., "AcceptabilityJudgment"). Check the \href{https://github.com/addrummond/ibex/blob/master/docs/manual.md}{IBEX documentation} for more details.
}
\description{
Reads in the \code{results} file from an IBEX experiment as data frame
}
