% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibex.R
\name{read_ibex}
\alias{read_ibex}
\title{Parse IBEX results file in Tidy format}
\usage{
read_ibex(results_file, clipboard = FALSE)
}
\arguments{
\item{results_file}{Path to the downloaded results file.}

\item{clipboard}{Read results from clipboard? If \code{TRUE}, do not need to supply argument for \code{results_file}.}
}
\value{
A single data frame containing all variables present as columns
}
\description{
Parse IBEX results file in Tidy format
}
