% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate_helpers.R
\name{split_track}
\alias{split_track}
\title{Split and track trajectory of individual levels of a group}
\usage{
split_track(df, grouping_var, tracked_along, ..., tracked_groups = NULL)
}
\arguments{
\item{df}{data frame}

\item{grouping_var}{column of the grouping variable used to split the data}

\item{tracked_along}{column of the tracking dimension to be passed into}

\item{...}{specification for the tracking window passed to dplyr::filter()}

\item{tracked_groups}{levels of the grouping_var to be tracked. Defaults to value NULL, which tracks all groups.}
}
\value{
a data frame
}
\description{
To be used with transition_reveal()
}
\examples{
\dontrun{
set.seed(123)

plot_df <- tibble(
  a = cumsum(runif(100, -1, 1)),
  b = cumsum(runif(100, -1.5, 1.5)),
  c = cumsum(runif(100, -2, 2)),
  d = cumsum(runif(100, -2.5, 2.5))
) \%>\% pivot_longer(a:d) \%>\%
  group_by(name) \%>\%
  mutate(reveal_time = row_number(), x = row_number()) \%>\%
  ungroup()

anim <- plot_df \%>\%
  split_track(grouping_var = name, tracked_along = reveal_time, x > 74) \%>\%
  ggplot(aes(x, value, color = name)) +
  geom_line(size = 1) +
  transition_reveal(reveal_time) +
  theme_classic()

animate(anim, width = 5, height = 3, units = "in", res = 150)
}
}
