% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2.R
\name{str2}
\alias{str2}
\title{Prints structure of an object recursively}
\usage{
str2(obj)
}
\arguments{
\item{obj}{A list or dataframe}
}
\value{
Prints to terminal
}
\description{
From stackoverflow answer - https://stackoverflow.com/questions/62970933/how-to-output-tree-like-human-readable-object-structure-in-r
}
\details{
For printing the structure of dataframes with list columns, it's useful to pass in just the first row if you know the elements of the list column are of the same type
}
\examples{
str2(list("Hello", list(", ", list("world!"))))
}
