% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{theme_junebug_calm}
\alias{theme_junebug_calm}
\title{ggplot2 theme inspired by \code{theme_gray()}}
\usage{
theme_junebug_calm(base_size = 12, ...)
}
\arguments{
\item{base_size}{Defaults to 12 (like all complete themes in ggplot2)}

\item{...}{Passed to \code{ggplot2::theme()}}
}
\description{
Features: Blue background, light grid lines
}
