% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{theme_junebug_clean}
\alias{theme_junebug_clean}
\title{ggplot2 theme inspired by \code{theme_classic()}}
\usage{
theme_junebug_clean(base_size = 12, ...)
}
\arguments{
\item{base_size}{Defaults to 12 (like all complete themes in ggplot2)}

\item{...}{Passed to \code{ggplot2::theme()}}
}
\description{
Features: Clear white background, no grid lines
}
