% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_toolbox.R
\name{tokens_flatten}
\alias{tokens_flatten}
\title{Reconstruct utterances from observations of tokens in tidy format}
\usage{
tokens_flatten(data, token_col, ...)
}
\arguments{
\item{data}{A data frame of tokens}

\item{token_col}{Name of column of tokens}

\item{...}{Grouping variables}
}
\value{
A data frame of grouping variables and utterances
}
\description{
Reconstruct utterances from observations of tokens in tidy format
}
\examples{

}
