% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_dummy.R
\name{col_dummy}
\alias{col_dummy}
\title{Apply explicit dummy coding to a data frame}
\usage{
col_dummy(data, factor, reference = NULL)
}
\arguments{
\item{data}{A data frame to be passed in (e.g., via piping)}

\item{factor}{A name of a factor column. Can be quoted or unquoted}

\item{reference}{A string of the level that should be coded as the reference level. Defaults to the first level of the factor if one is not provided.}
}
\value{
A \code{data.frame}. A tibble passed into \code{data} is returned as a tibble.
}
\description{
Apply explicit dummy coding to a data frame
}
\examples{
# column name can be quoted or unquoted
col_dummy(iris, "Species", reference = "virginica")
col_dummy(iris, Species, reference = "virginica")

# model interpretation is the same
lm(Sepal.Length ~ Species, data = iris)
lm(Sepal.Length ~ Speciesversicolor + Speciesvirginica, data = col_dummy(iris, Species))
}
