% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{icon_html}
\alias{icon_html}
\title{Find a Font Awesome icon in the appropriate html format}
\usage{
icon_html(icon, type = "solid", unicode = TRUE)
}
\arguments{
\item{icon}{A character string that matches one or more Font Awesome icon names}

\item{type}{One of "brand", "regular", and "solid"}

\item{unicode}{Should it return the unicode value for the icon wrapped in <span>? If \code{FALSE}, returns the icon class wrapped in <i>}
}
\value{
A character vector of matches in the specified format
}
\description{
Find a Font Awesome icon in the appropriate html format
}
