% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibex.R
\name{read_ibex}
\alias{read_ibex}
\title{Reads in the \code{results} file from an IBEX experiment as data frame}
\usage{
read_ibex(results_file, external = TRUE)
}
\arguments{
\item{results_file}{The exact text of the \code{results} file as downloaded from IBEX}

\item{external}{Is the results file a path to a downloaded file? Set to \code{FALSE} if the results are formatted as lines of characters.}
}
\value{
A data frame if the controller writes a single line to the results file (e.g., "DashedSentence") OR a list of data frames if the controller writes more than one line to the results file (e.g., "AcceptabilityJudgment"). Check the \href{https://github.com/addrummond/ibex/blob/master/docs/manual.md}{IBEX documentation} for more details.
}
\description{
Reads in the \code{results} file from an IBEX experiment as data frame
}
